requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0